Copying style folder for local theme
Adding info about copying the styles folder into the folder of the exported file. This works if exporting a few files residing in a few folders but very tedious for a large number of folders (as is in my case).

I tried to adjust the href in the .setup file to make it an absolute path with ~ for home directory, it did not work however. suggestions appreciated.